+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.2.0 (unreleased)
 
-- Added support for `sparsevec` type
+- Added support for `halfvec` and `sparsevec` type
 
 ## 0.1.1 (2023-03-09)